Semantic_Analyzer now compiles (with warnings)
I went through and fixed the compilation errors. Some problematic code
has been commented out; that will be completed ASAP.

Changes to SAST:
1. Varidentifier because I don't know what to assign to identifiers. We
already have an identifier type, so Varidentifier seems like a good type
to represent it in SAST.

2. Sorry John, I personally believe that operator should be kept in AST
only and imported to. Whoever feels this is not the case, try debugging
the binops in the semantic analyzer and let me know if you can get it
working by not opening the AST operators in SAST.